autoflow/requirements.txt
pandas==2.2.2
openpyxl==3.1.5
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
playwright==1.46.0
pytest==8.3.2
